Now with Array of classes I can have a Player class, with Player 1 and Player 2 as objects In the Player class there is a PkmnList array with accepts Pkmn objects There I can store the Pokemon In Main I will creat a new Pkmn and tghen store it in P1 or P2's Pkmn List And in the battle menu i sent P1 and P2 classes and then create a vatabile with Pkmn object named pokemon1 and pokemon2 who will copy the current pokemon. When im done wih pokemon1 and pokemon2 i can just use the getPknlist module and it will store the newc hanges in data of the pokemon Thus the program backbone is almsot complete
--Commit 2
So I created player class. Now how to deal with Pkmn Moves
So I create 2 Player objects.
Then I create a Pkmn object that will go into a Player's pkmnlist array
Pkmn has a arraya PkmnMovelist with holds PkmnMove classes or otherwords this array holds the Pkmn's moves
when starting Pkmn's builder it has a code that will ask for species, and that will use the PkmnSpecies Class
In this species class, there will be a if statements that will give the Pkmn the moveset it needs according to which species is passed
When it finds the right species, to set the moves it goes to the Pokemon Move Registry Class, sending it the move name
PPjmnMoveRegistry class asks for a move name and it will have if statements for each movename. It creates a PkmnMove object and returns.
Back to Pkmn Species. Pkmn Species then adds that move to the MoveList array, thus Pkmn will have its movelist at end of intaliziatiojn.

Currently testing if this whole process works.